Some improvements
in css class it will be checked if the css file exists if yes it will check if it empty or not after that the css rules will be written to css file and if or the new css rules will be append to the css file if it is  not empty.

+++ Fonts Downloader/CssFile.cs
@@ -17,15 +17,18 @@ namespace Fonts_Downloader
             { "900", "Black" },
         };
 
-        public void CreateCSS(IEnumerable<string> variants, string folderName, string fontName, bool woff, bool ttf, bool minify = false, IEnumerable<string> subsets = null)
+        public void CreateCSS(IEnumerable<string> variants, string folderName, string fontName, bool woff, bool minify = false, IEnumerable<string> subsets = null)
         {
 
-            var cssList = GenerateCssList(variants, fontName, woff, ttf, subsets);
+            var cssList = GenerateCssList(variants, fontName, woff, subsets);
 
             if (cssList.Any())
             {
                 string fontFolder = Path.Combine(folderName, fontName.Replace(" ", ""));
-                Directory.CreateDirectory(fontFolder);
+                if (!Directory.Exists(fontFolder))
+                {
+                    Directory.CreateDirectory(fontFolder);
+                }
 
                 string cssFilePath = Path.Combine(fontFolder, $"{fontName.Replace(" ", "")}{(minify ? ".min" : "")}.css");
                 string cssContent = string.Join("\n", cssList);
@@ -35,12 +38,31 @@ namespace Fonts_Downloader
                     var minifier = new Minifier();
                     cssContent = minifier.MinifyStyleSheet(cssContent);
                 }
+                if (!File.Exists(cssFilePath) || (File.Exists(cssFilePath) && string.IsNullOrEmpty(File.ReadAllText(cssFilePath))))
+                    File.WriteAllText(cssFilePath, cssContent);
 
-                File.WriteAllText(cssFilePath, cssContent);
+                else
+                {
+                    string fileContents = File.ReadAllText(cssFilePath);  
+                    if (woff && fileContents.Contains(".ttf") || !woff && fileContents.Contains(".woff2"))
+                    {
+                        File.WriteAllText(cssFilePath, cssContent);
+                    }
+                    else
+                    {
+                        if (!fileContents.Contains(cssContent))
+                        {
+                            string newString = cssContent.Replace(fileContents, "");
+                            int index = fileContents.IndexOf(cssContent);
+                            if (index == -1)
+                                File.AppendAllText(cssFilePath, newString);
+                        }
+                    }
+                }
             }
         }
 
-        private List<string> GenerateCssList(IEnumerable<string> variants, string fontName, bool woff, bool ttf, IEnumerable<string> subsets = null)
+        private List<string> GenerateCssList(IEnumerable<string> variants, string fontName, bool woff, IEnumerable<string> subsets = null)
         {
             var cssList = new List<string>();
 
@@ -100,7 +122,7 @@ namespace Fonts_Downloader
                 css += "  unicode-range: U+0460-052F, U+1C80-1C88, U+20B4, U+2DE0-2DFF, U+A640-A69F, U+FE2E-FE2F;\n";
             }
 
-            css += "}\n";
+            css += "}";
 
             return css;
         }
